requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
configparser==3.7.4
contextlib2==0.5.5
Flask==1.1.0
funcsigs==1.0.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
jsonify==0.5
MarkupSafe==1.1.1
more-itertools==5.0.0
mysqlclient==1.4.2.post1
packaging==19.0
pathlib2==2.3.4
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.4
python-dateutil==2.8.0
requests==2.22.0
scandir==1.10.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
